requirements.txt
streamlit>=1.42.0
anthropic>=0.94.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
finvizfinance>=0.14.0
pytest>=7.0.0
